Moved list queries into sub module
