+++ app/graphql/queries/list/base.rb
@@ -1,0 +1,14 @@
+module Queries
+  module List
+    class Base < GraphQL::Function
+      def call(_obj, _args, ctx)
+        ctx[:pundit].authorize model_name.constantize, :index?
+        yield
+      end
+
+      def model_name
+        self.class.name.split('List').last.singularize
+      end
+    end
+  end
+end
