--- app/graphql/queries/list_record_books.rb
@@ -1,12 +1,0 @@
-module Queries
-  class ListRecordBooks < ListRecords
-    description 'List record books with various filters'
-    type types[::Types::RecordBookType]
-
-    def call(_obj, _args, ctx)
-      super do
-        ctx[:pundit].policy_scope RecordBook.all
-      end
-    end
-  end
-end
