--- app/graphql/queries/list_teams.rb
@@ -1,12 +1,0 @@
-module Queries
-  class ListTeams < ListRecords
-    description 'List teams with various filters'
-    type types[::Types::TeamType]
-
-    def call(_obj, _args, ctx)
-      super do
-        ctx[:pundit].policy_scope Team.all
-      end
-    end
-  end
-end
